refactor: move scan.py and its tests into codecity package
scan.py → codecity/scan.py; test_scan.py → codecity/tests/test_scan.py
(now relies on the installed package, no sys.path hack); fixture setup
moved alongside its test. Drop the unittest-discover docstring — pytest
auto-collects unittest.TestCase classes.
