+++ .gitignore
@@ -6,8 +6,8 @@
 docs/superpowers/
 
 # Generated test fixtures
-#   run `bash tests/fixtures/setup.sh` to regenerate
-tests/fixtures/sample-repo/
+#   run `bash codecity/tests/fixtures/setup.sh` to regenerate
+codecity/tests/fixtures/sample-repo/
 
 # Dev-mode scratch (codecity.py --dev writes manifest/config here)
 build/
